requirements.txt
aiogram>=3.4.0
langchain>=0.2.0
qdrant-client>=1.9.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
mcp
websockets>=12.0
